+++ src/Auth/Implement/JwtWithPrivateKeyAuthentication.php
@@ -29,7 +29,7 @@ class JwtWithPrivateKeyAuthentication extends \YdbPlatform\Ydb\Auth\JwtAuth
             'jwt' => $jwt_token,
         ];
         $token = $this->requestToken($request_data);
-        return new TokenInfo($token->iamToken, $this->convertExpiresAt($token->expiresAt));
+        return new TokenInfo($token->iamToken, $this->convertExpiresAt($token->expiresAt), $this->refreshTokenRatio);
     }
 
     public function getName(): string