+++ src/Auth/Implement/OAuthTokenAuthentication.php
@@ -22,7 +22,7 @@ class OAuthTokenAuthentication extends \YdbPlatform\Ydb\Auth\IamAuth
             'yandexPassportOauthToken' => $this->oauth_token,
         ];
         $token = $this->requestToken($request_data);
-        return new TokenInfo($token->iamToken, $this->convertExpiresAt($token->expiresAt));
+        return new TokenInfo($token->iamToken, $this->convertExpiresAt($token->expiresAt), $this->refreshTokenRatio);
     }
 
     public function getName(): string